Minor tweaks

+++ Texture.h
@@ -33,7 +33,7 @@ namespace ObjMaster {
 		 * It has the special value of 0 when the texture is not loaded
 		 * onto the GPU unit.
 		 */
-		unsigned int handle{};
+		uintptr_t handle{};
 		/** Width in pixels */
 		int width{};
 		/** Height in pixels */
